+++ P1G1Shooter/P1G1Shooter/projectile.h
@@ -17,7 +17,7 @@ typedef struct Projectile
 /// <param name="proj">Projectile to initialize</param>
 /// <param name="speed">Speed of the projectile</param>
 /// <param name="direction">Direction of the projectile, either left or right</param>
-void Proj_Initialize(Projectile** proj, float speed, int health,
+void Proj_Initialize(Projectile** proj, float speed, float health,
 	float velocity_x, float velocity_y,
 	double pos_x, double pos_y, 
 	EntityType type, GameScreenData* gameScreen,
